Bug: SF-115 Rebase changes after identity cleanup
Fix it to actually work!
