--- src/SIL.XForge.Identity/Models/VerifyInviteUserParams.cs
@@ -1,7 +1,0 @@
-namespace SIL.XForge.Identity.Models
-{
-    public class VerifyInviteUserParams
-    {
-        public string Email { get; set; }
-    }
-}
