+++ src/SIL.XForge.Scripture/ClientApp/src/identity/identity.service.ts
@@ -28,17 +28,16 @@ export class IdentityService {
     return this.invoke('signUp', name, password, email, recaptcha);
   }
 
+  verifyInvitedUser(email: string): Promise<boolean> {
+    return this.invoke('verifyInvitedUser', email);
+  }
+
   verifyEmail(key: string): Promise<boolean> {
     return this.invoke('verifyEmail', key);
   }
 
   verifyResetPasswordKey(key: string): Promise<boolean> {
     return this.invoke('verifyResetPasswordKey', key);
-  }
-
-  async verifyInvitedUser(email: string): Promise<boolean> {
-    const result = await this.callApi('verify-invited-user', { email } as VerifyInvitedUserParams);
-    return result.success;
   }
 
   private async invoke<T>(method: string, ...params: any[]): Promise<T> {
