--- src/SIL.XForge.Scripture/ClientApp/src/identity/models/verify-invited-user-params.ts
@@ -1,3 +1,0 @@
-export interface VerifyInvitedUserParams {
-  email: string;
-}
